+++ badminton-api/src/main/java/org/badminton/api/club/service/ClubAddService.java
@@ -2,7 +2,7 @@ package org.badminton.api.club.service;
 
 import org.badminton.api.club.model.dto.ClubAddRequest;
 import org.badminton.api.club.model.dto.ClubAddResponse;
-import org.badminton.api.club.provider.ClubDataProvider;
+import org.badminton.api.club.validator.ClubValidator;
 import org.badminton.domain.club.entity.ClubEntity;
 import org.springframework.stereotype.Service;
 
@@ -14,7 +14,7 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor
 public class ClubAddService {
 
-	private final ClubDataProvider clubDataProvider;
+	private final ClubValidator clubDataProvider;
 
 	// TODO: clubAddRequest에 이미지가 없으면 default 이미지를 넣어주도록 구현
 	public ClubAddResponse addClub(ClubAddRequest clubAddRequest) {
@@ -22,7 +22,7 @@ public class ClubAddService {
 		ClubEntity club = new ClubEntity(clubAddRequest.getClubName(), clubAddRequest.getClubDescription(),
 			clubAddRequest.getClubImage());
 		clubDataProvider.saveClub(club);
-		return ClubAddResponse.toClubAddResponse(club);
+		return ClubAddResponse.clubEntityToClubAddResponse(club);
 	}
 
 }
